--- client/backlog/backlog.html
@@ -1,15 +1,0 @@
-<template name="backlog">
-  <body>
-    <h3>Backlog</h3>
-
-    <form class="new-task">
-      <input type="text" name="text" placeholder="Task Description" />
-    </form>
-
-    <ul>
-      {{#each singleTask in tasks}}
-        {{> task task=singleTask}}
-      {{/each}}
-    </ul>
-  </body>
-</template>
