--- client/main.css
@@ -1,1 +1,0 @@
-/* CSS declarations go here */
