+++ client/main.js
@@ -1,4 +1,1 @@
-import { Template } from 'meteor/templating';
-import { ReactiveVar } from 'meteor/reactive-var';
-
-import './main.html';
+import '/imports/startup/client/index.js';
