+++ imports/api/api.js
@@ -1,0 +1,17 @@
+import { Tasks } from '/tasks/tasks.js'
+import '/tasks/methods.js'
+
+if (Meteor.isServer) {
+  // This code only runs on the server
+  // Only publish tasks that are public or belong to the current user
+  Meteor.publish('tasks', function tasksPublication() {
+    return Tasks.find({
+      $or: [
+        { private: { $ne: true } },
+        { owner: this.userId },
+      ],
+    });
+  });
+}
+
+export { Tasks }
