+++ imports/api/tasks/tasks.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const Tasks = new Mongo.Collection('tasks');
