+++ imports/startup/client/index.js
@@ -1,0 +1,2 @@
+// import './useraccounts-configuration.js';
+import './routes.js';
