+++ imports/ui/layouts/layout.html
@@ -1,4 +1,4 @@
-<template name="example">
+<template name="layout">
   <head>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
@@ -8,8 +8,7 @@
     <meta name="author" content="">
     <link rel="icon" href="../../favicon.ico">
 
-    <title>Dashboard</title>
-
+    <title>CMU-DO</title>
   </head>
 
   <body>
@@ -35,14 +34,7 @@
       </div>
     </nav>
 
-    <div class="container">
-
-      <div class="starter-template">
-        <h1>Bootstrap starter template</h1>
-        <p class="lead">Use this document as a way to quickly start any new project.<br> All you get is this text and a mostly barebones HTML document.</p>
-      </div>
-
-    </div><!-- /.container -->
+    {{> Template.dynamic template=page}}
 
   </body>
 </template>
