+++ imports/ui/pages/backlog.html
@@ -1,0 +1,13 @@
+<template name="backlog">
+  <h3>Backlog</h3>
+
+  <form class="new-task">
+    <input type="text" name="text" placeholder="Task Description" />
+  </form>
+
+  <ul>
+    {{#each singleTask in tasks}}
+      {{> task task=singleTask}}
+    {{/each}}
+  </ul>
+</template>
