+++ server/main.js
@@ -1,6 +1,1 @@
-import { Meteor } from 'meteor/meteor';
-import '../imports/api/tasks.js';
-
-Meteor.startup(() => {
-  // code to run on server at startup
-});
+import '/imports/startup/server/index.js';
